refactor(utils): move toPascalCase to sync-figma instead of utils-shared
