+++ .nx/version-plans/version-plan-1788534812000-ui-react.md
@@ -1,0 +1,5 @@
+---
+'@ledgerhq/lumen-ui-react': patch
+---
+
+fix(Icon): use the icon name as displayName without PascalCase conversion
