+++ .nx/version-plans/version-plan-1788534967000-utils-shared.md
@@ -1,0 +1,5 @@
+---
+'@ledgerhq/lumen-utils-shared': patch
+---
+
+Remove the unused `toPascalCase` helper from the public API.
